Fixed task related to cms account delete

+++ app/jobs/users/cancel_user_remote_domain_job.rb
@@ -4,7 +4,7 @@ module Users
 
     def perform(user_id:)
       user  = User.find user_id
-      if !user.api_id.nil? and user.club.billing_enable
+      if user.api_user and !user.api_id.nil? and user.club.billing_enable
         res = user.api_user.destroy! 
         raise 'CancelUserRemoteDomainJob::UnexpectedError: CMS Account not destroyed.' if res.nil?
         raise res.body['error_message'] if !res.body['success']
